requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
dj-database-url==2.3.0
dj-rest-auth==5.0.1
Django==4.2.18
django-allauth==65.3.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
graphviz==0.20.3
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
